Split out effects.rs
