+++ crates/sui-benchmark/src/lib.rs
@@ -30,9 +30,9 @@ use sui_json_rpc_types::{
 };
 use sui_network::{DEFAULT_CONNECT_TIMEOUT_SEC, DEFAULT_REQUEST_TIMEOUT_SEC};
 use sui_sdk::{SuiClient, SuiClientBuilder};
+use sui_types::effects::{CertifiedTransactionEffects, TransactionEffectsAPI, TransactionEvents};
 use sui_types::messages::CallArg;
 use sui_types::messages::ObjectArg;
-use sui_types::messages::TransactionEvents;
 use sui_types::programmable_transaction_builder::ProgrammableTransactionBuilder;
 use sui_types::sui_system_state::sui_system_state_summary::SuiSystemStateSummary;
 use sui_types::{
@@ -43,10 +43,7 @@ use sui_types::{
         AuthoritySignature,
     },
     message_envelope::Envelope,
-    messages::{
-        CertifiedTransaction, CertifiedTransactionEffects, HandleCertificateResponse, Transaction,
-        TransactionEffectsAPI, TransactionStatus,
-    },
+    messages::{CertifiedTransaction, HandleCertificateResponse, Transaction, TransactionStatus},
     object::Object,
 };
 use sui_types::{base_types::ObjectRef, crypto::AuthorityStrongQuorumSignInfo, object::Owner};
