+++ crates/sui-core/src/unit_tests/gas_tests.rs
@@ -12,6 +12,7 @@ use move_core_types::ident_str;
 use once_cell::sync::Lazy;
 use sui_protocol_config::ProtocolConfig;
 use sui_types::crypto::AccountKeyPair;
+use sui_types::effects::TransactionEvents;
 use sui_types::execution_status::{ExecutionFailureStatus, ExecutionStatus};
 use sui_types::gas::SuiCostTable;
 use sui_types::gas_coin::GasCoin;
