+++ crates/sui-core/src/unit_tests/move_package_upgrade_tests.rs
@@ -7,10 +7,7 @@ use sui_protocol_config::ProtocolConfig;
 use sui_types::{
     base_types::{ObjectID, ObjectRef, SuiAddress},
     crypto::{get_key_pair, AccountKeyPair},
-    messages::{
-        Argument, ObjectArg, ProgrammableTransaction, TransactionEffects, TransactionEffectsV1,
-        TEST_ONLY_GAS_UNIT_FOR_PUBLISH,
-    },
+    messages::{Argument, ObjectArg, ProgrammableTransaction, TEST_ONLY_GAS_UNIT_FOR_PUBLISH},
     move_package::UpgradePolicy,
     object::{Object, Owner},
     programmable_transaction_builder::ProgrammableTransactionBuilder,
@@ -19,6 +16,7 @@ use sui_types::{
 };
 
 use std::{collections::BTreeSet, path::PathBuf, str::FromStr, sync::Arc};
+use sui_types::effects::{TransactionEffects, TransactionEffectsV1};
 use sui_types::execution_status::{
     CommandArgumentError, ExecutionFailureStatus, PackageUpgradeError,
 };
