+++ crates/sui-json-rpc-types/src/sui_transaction.rs
@@ -23,14 +23,14 @@ use sui_types::base_types::{
     EpochId, ObjectID, ObjectRef, SequenceNumber, SuiAddress, TransactionDigest,
 };
 use sui_types::digests::{ObjectDigest, TransactionEventsDigest};
+use sui_types::effects::{TransactionEffects, TransactionEffectsAPI, TransactionEvents};
 use sui_types::error::{ExecutionError, SuiError};
 use sui_types::execution_status::ExecutionStatus;
 use sui_types::gas::GasCostSummary;
 use sui_types::messages::{
     Argument, CallArg, Command, ExecuteTransactionRequestType, GenesisObject, InputObjectKind,
     ObjectArg, ProgrammableMoveCall, ProgrammableTransaction, SenderSignedData, TransactionData,
-    TransactionDataAPI, TransactionEffects, TransactionEffectsAPI, TransactionEvents,
-    TransactionKind, VersionedProtocolMessage,
+    TransactionDataAPI, TransactionKind, VersionedProtocolMessage,
 };
 use sui_types::messages_checkpoint::CheckpointSequenceNumber;
 use sui_types::object::Owner;
