+++ crates/sui-json-rpc/src/transaction_execution_api.rs
@@ -21,9 +21,9 @@ use sui_json_rpc_types::{
 };
 use sui_open_rpc::Module;
 use sui_types::base_types::SuiAddress;
+use sui_types::effects::TransactionEffectsAPI;
 use sui_types::messages::{
-    ExecuteTransactionRequest, ExecuteTransactionRequestType, TransactionEffectsAPI,
-    TransactionKind,
+    ExecuteTransactionRequest, ExecuteTransactionRequestType, TransactionKind,
 };
 use sui_types::messages::{ExecuteTransactionResponse, Transaction};
 use sui_types::messages::{TransactionData, TransactionDataAPI};
