+++ crates/sui-types/src/messages_checkpoint.rs
@@ -8,10 +8,10 @@ use crate::crypto::{
     default_hash, AggregateAuthoritySignature, AuthoritySignInfo, AuthorityStrongQuorumSignInfo,
 };
 use crate::digests::Digest;
+use crate::effects::TransactionEffectsAPI;
 use crate::error::SuiResult;
 use crate::gas::GasCostSummary;
 use crate::message_envelope::{Envelope, Message, TrustedEnvelope, VerifiedEnvelope};
-use crate::messages::TransactionEffectsAPI;
 use crate::signature::GenericSignature;
 use crate::storage::ReadStore;
 use crate::sui_serde::AsProtocolVersion;
